refactor: reorganize lib into blog/ and seo/ directories
Move blog loading, frontmatter validation, and SEO helpers out of flat
lib/ files into domain-specific folders and split site config from
page metadata builders.
