+++ app/blog/[slug]/opengraph-image.tsx
@@ -1,11 +1,11 @@
 import { notFound } from "next/navigation";
+import { getPostFrontmatterBySlug } from "@/lib/blog/frontmatter";
+import { getAllSlugs } from "@/lib/blog/posts";
 import {
 	createOgImageResponse,
 	ogImageContentType,
 	ogImageSize,
 } from "@/lib/og/image-response";
-import { getPostFrontmatterBySlug } from "@/lib/post-frontmatter";
-import { getAllSlugs } from "@/lib/posts";
 
 export const size = ogImageSize;
 export const contentType = ogImageContentType;
