+++ mdx.d.ts
@@ -1,6 +1,6 @@
 declare module "*.mdx" {
 	import type { ComponentType } from "react";
-	import type { BaseFrontmatter } from "./lib/posts";
+	import type { BaseFrontmatter } from "./lib/blog/posts";
 
 	const MDXComponent: ComponentType;
 	export const frontmatter: BaseFrontmatter;
